requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
fastapi==0.115.12
    # via solview (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.71.0
    # via opentelemetry-exporter-otlp-proto-grpc
idna==3.10
    # via
    #   anyio
    #   requests
importlib-metadata==8.6.1
    # via opentelemetry-api
loguru==0.7.3
    # via solview (pyproject.toml)
opentelemetry-api==1.32.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   solview (pyproject.toml)
opentelemetry-exporter-otlp==1.32.1
    # via solview (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.32.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.32.1
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.32.1
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.53b1
    # via solview (pyproject.toml)
opentelemetry-proto==1.32.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.32.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   solview (pyproject.toml)
opentelemetry-semantic-conventions==0.53b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
orjson==3.10.18
    # via solview (pyproject.toml)
packaging==25.0
    # via opentelemetry-instrumentation
prometheus-client==0.21.1
    # via solview (pyproject.toml)
protobuf==5.29.4
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.11.4
    # via
    #   fastapi
    #   solview (pyproject.toml)
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.0
    # via solview (pyproject.toml)
requests==2.32.3
    # via opentelemetry-exporter-otlp-proto-http
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via
    #   fastapi
    #   solview (pyproject.toml)
typing-extensions==4.13.2
    # via
    #   fastapi
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.21.0
    # via importlib-metadata
